backend/requirements.txt
google-adk
fastapi
uvicorn[standard]
websockets
python-dotenv
asyncio
pydantic